+++ python/ffsim/qiskit/gates/ucj_operator.py
